+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## Unreleased
-- Improved README
+- Add SEV (`--mode=sev`) and SEV-ES (`--mode=seves`) measurement modes
+- Add `--verbose` for verbose output; by default, only the digest is printed
+- Improve README
 
 ## 0.0.2 - 2022-04-11